+++ app/package.json
@@ -15,6 +15,7 @@
     "@antv/data-set": "^0.10.1",
     "ant-design-vue": "~1.3.8",
     "axios": "^0.18.0",
+    "core-js": "^2.6.5",
     "enquire.js": "^2.1.6",
     "install": "^0.12.2",
     "js-cookie": "^2.2.0",
